+++ graphql-gradle-plugin/src/main/java/com/graphql_java_generator/gradleplugin/GenerateGraphQLSchemaSpringConfiguration.java
@@ -14,7 +14,9 @@ import org.springframework.context.annotation.ComponentScan.Filter;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.FilterType;
 
+import com.graphql_java_generator.plugin.DocumentParser;
 import com.graphql_java_generator.plugin.ResourceSchemaStringProvider;
+import com.graphql_java_generator.util.GraphqlUtils;
 
 import graphql.language.Document;
 import graphql.parser.Parser;
@@ -24,7 +26,7 @@ import graphql.parser.Parser;
  */
 @Configuration
 // @Import({ JacksonAutoConfiguration.class })
-@ComponentScan(basePackages = { "com.graphql_java_generator" }, excludeFilters = {
+@ComponentScan(basePackageClasses = { DocumentParser.class, GraphqlUtils.class }, excludeFilters = {
 		@Filter(type = FilterType.REGEX, pattern = ".*\\.GenerateClientCode.*"),
 		@Filter(type = FilterType.REGEX, pattern = ".*\\.GenerateCode.*"),
 		@Filter(type = FilterType.REGEX, pattern = ".*\\.GeneratePojo.*"),