+++ CHANGELOG.md
@@ -6,6 +6,14 @@ Whether the application uses the _graphql_, the _generateClientCode_ or the _gen
 * generateBatchLoaderEnvironment: true _(server only)_
 * generateDeprecatedRequestResponse: false _(client only)_
 * separateUtilityClasses: true _(both client and server mode)_
+* skipGenerationIfSchemaHasNotChanged: true _(both client and server mode)_
+
+# 1.15
+
+Client mode:
+* The queries, mutations and subscriptions now __accept GraphQL aliases__, like specified in [GraphQL spec](http://spec.graphql.org/June2018/#sec-Field-Alias)
+    * Once the request is executed, The alias value can be retrieved with the `Object getAliasValue(String: aliasname)` method that has been added to every generated objects and interfaces. This method returns the alias value, parsed into the relevant Java Object.
+* The default `QueryExecutor` provided by the plugin is now a Spring bean. If you want to override it, you should not mark your own `QueryExecutor` with the `@Primary` annotation, to ignore the default one.
 
 
 # 1.14.2